# Store finalizedEpoch when processing registry updates