Another step in reducing state accesses.

```
before:
Benchmark                                 (validatorsCount)   Mode  Cnt  Score   Error  Units
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.470 ± 0.051  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.557 ± 0.076  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.523 ± 0.035  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.542 ± 0.053  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.499 ± 0.058  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.548 ± 0.061  ops/s

after
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.534 ± 0.039  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.555 ± 0.065  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.467 ± 0.031  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.505 ± 0.127  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.558 ± 0.054  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.548 ± 0.049  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.522 ± 0.033  ops/s
```

Doesn't appear to speed things up but is in line with our general direction.

Signed-off-by: Paul Harris <paul.harris@consensys.net>

## Documentation

- [X] I thought about documentation and added the `doc-change-required` label to this PR if updates are required.

## Changelog

- [X] I thought about adding a changelog entry, and added one if I deemed necessary.